requirements.txt
playwright>=1.40.0

pyautogui>=0.9.54
pillow>=10.0.0
opencv-python>=4.8.0